app/api/permissions.py
from rest_framework.permissions import BasePermission


class IsOwnerOnly(BasePermission):
    """
    Allows access only to authenticated users.
    """

    def has_permission(self, request, view):
        return request.user and request.user.is_authenticated

    def has_object_permission(self, request, view, obj):
        return request.user and request.user.is_authenticated and request.user == obj.user
